+++ src/AppkeepService.php
@@ -3,6 +3,7 @@
 namespace Appkeep\Laravel;
 
 use InvalidArgumentException;
+use Appkeep\Laravel\Health\Check;
 
 class AppkeepService
 {
@@ -22,14 +23,14 @@ class AppkeepService
 
     public function checks(array $checks = [], $replace = false)
     {
-        if (!app()->runningInConsole()) {
+        if (! app()->runningInConsole()) {
             return;
         }
 
         foreach ($checks as $check) {
             $this->rejectIfDoesNotExtendBaseClass($check);
 
-            if (!$replace) {
+            if (! $replace) {
                 $this->rejectIfDuplicate($check);
             }
 
@@ -39,7 +40,6 @@ class AppkeepService
         return collect($this->checks);
     }
 
-
     public function replaceChecks(array $checks = [])
     {
         return $this->checks($checks, true);
@@ -47,7 +47,7 @@ class AppkeepService
 
     protected function rejectIfDoesNotExtendBaseClass($check)
     {
-        if (!($check instanceof Check)) {
+        if (! ($check instanceof Check)) {
             throw new InvalidArgumentException(
                 sprintf('%s is not an instance of %s', get_class($check), Check::class)
             );
